msg/async: msgr2: extract frame classes into its own header
